prevent deleting own account

+++ src/main/java/io/muzoo/ooc/webapp/basic/servlets/DeleteUserServlet.java
@@ -3,6 +3,7 @@ package io.muzoo.ooc.webapp.basic.servlets;
 import io.muzoo.ooc.webapp.basic.security.SecurityService;
 import io.muzoo.ooc.webapp.basic.security.User;
 import io.muzoo.ooc.webapp.basic.security.UserService;
+import org.apache.commons.lang.StringUtils;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -25,12 +26,19 @@ public class DeleteUserServlet extends AbstractRoutableHttpServlet {
             try{
                 User currentUser = userService.findByUsername(username);
                 User deletingUser = userService.findByUsername(request.getParameter("username"));
-                if(userService.deleteUserByName(deletingUser.getUsername())){
-                    request.getSession().setAttribute("hasError",false);
-                    request.getSession().setAttribute("message",String.format("User %s is successfully deleted.",deletingUser.getUsername()));
-                }else{
+
+                if(StringUtils.equals(currentUser.getUsername(), deletingUser.getUsername())) {
                     request.getSession().setAttribute("hasError",true);
-                    request.getSession().setAttribute("message",String.format("Unable to delete user %s.",deletingUser.getUsername()));
+                    request.getSession().setAttribute("message",String.format("you cannot delete your own account."));
+                }
+                else{
+                    if(userService.deleteUserByName(deletingUser.getUsername())){
+                        request.getSession().setAttribute("hasError",false);
+                        request.getSession().setAttribute("message",String.format("User %s is successfully deleted.",deletingUser.getUsername()));
+                    }else{
+                        request.getSession().setAttribute("hasError",true);
+                        request.getSession().setAttribute("message",String.format("Unable to delete user %s.",deletingUser.getUsername()));
+                    }
                 }
             }
             catch(Exception e){
